Add Requires for openstack-puppet-modules
Install the openstack-puppet-modules package via spec file Requires. This
allows us to remove the dependency on having a dib style element do it.

It's still valid to use instack-undercloud when using puppet modules from
source instead of packages. In that case, the package will still be installed
due to this change, but wouldn't have any effect on the puppet modules actually
in use since they are only installed under /usr/share and not under
/etc/puppet/modules.

Change-Id: Ida9e7509e8a823e51f051c82dc298cc554f4d71f